Move static constants from IFileStorage into CommonFileStorage
