+++ tracdap-plugins/aws-storage/src/main/java/org/finos/tracdap/plugins/aws/storage/S3ObjectReader.java
@@ -38,7 +38,7 @@ import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Flow;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import static org.finos.tracdap.common.storage.IFileStorage.READ_OPERATION;
+import static org.finos.tracdap.common.storage.CommonFileStorage.READ_OPERATION;
 import static org.finos.tracdap.common.storage.StorageErrors.ExplicitError.DUPLICATE_SUBSCRIPTION;
